+++ Model/Validate.php
@@ -3,8 +3,12 @@
 namespace OpenApi\Model;
 
 use HKarlstrom\Middleware\OpenApiValidation;
+use HKarlstrom\Middleware\OpenApiValidation\Exception\FileNotFoundException;
+use HKarlstrom\Middleware\OpenApiValidation\Exception\InvalidOptionException;
+use HKarlstrom\OpenApiReader\OpenApiReader;
 use MVC\Cache;
 use MVC\Config;
+use MVC\Convert;
 use MVC\DataType\DTRequestCurrent;
 use MVC\Debug;
 use MVC\Error;
@@ -94,11 +98,82 @@ class Validate
             return $oDTValidateRequestResponse;
         }
 
-        // requirement: it has to be Psr7
-        $oPsrRequest = new PsrRequest($oDTRequestCurrent);
+        // check the request content type
+        try {
+            $oOpenApiReader = new OpenApiReader($sYamlFileAbs);
+
+            // get the expected type of request
+            $sExpectedType = $oOpenApiReader->getOperationRequestBody(
+                $oDTRequestCurrent->get_path(),
+                strtolower($oDTRequestCurrent->get_requestmethod())
+            )->getContent()->type;
+
+            // check content type "json"
+            if (true === (boolean) stristr($sExpectedType, 'json') && false === Strings::isJson($oDTRequestCurrent->get_input()))
+            {
+                $sMessage = 'content type has to be valid `' . $sExpectedType . '`';
+                Error::error(json_last_error_msg() . ' on RequestBody of ' . $oDTRequestCurrent->get_path() . ': ' . $sMessage);
+                Error::notice('abort validation of request due to error');
+                $oDTValidateRequestResponse
+                    ->set_bSuccess(false)
+                    ->add_aMessage(
+                        DTValidateMessage::create()
+                            ->set_sSubject('Error')
+                            ->set_sBody(json_last_error_msg())
+                    )
+                    ->add_aMessage(
+                        DTValidateMessage::create()
+                            ->set_sSubject('Notice')
+                            ->set_sBody($sMessage)
+                    );
+
+                return $oDTValidateRequestResponse;
+            }
+        } catch (\Exception $oException) {
+            Error::exception($oException->getMessage());
+            Error::notice('abort validation of request due to exception');
+            $oDTValidateRequestResponse
+                ->set_bSuccess(false)
+                ->add_aMessage(
+                    DTValidateMessage::create()
+                        ->set_sSubject('Exception')
+                        ->set_sBody($oException->getMessage())
+                );
+
+            return $oDTValidateRequestResponse;
+        }
 
         // OpenApiValidation
-        $oOpenApiValidation = new OpenApiValidation($sYamlFileAbs);
+        try {
+            $oOpenApiValidation = new OpenApiValidation($sYamlFileAbs);
+        } catch (FileNotFoundException $oFileNotFoundException) {
+            Error::exception($oFileNotFoundException->getMessage());
+            Error::notice('abort validation of request due to exception');
+            $oDTValidateRequestResponse
+                ->set_bSuccess(false)
+                ->add_aMessage(
+                    DTValidateMessage::create()
+                        ->set_sSubject('Exception')
+                        ->set_sBody($oFileNotFoundException->getMessage())
+                );
+
+            return $oDTValidateRequestResponse;
+        } catch (InvalidOptionException $oInvalidOptionException) {
+            Error::exception($oInvalidOptionException->getMessage());
+            Error::notice('abort validation of request due to exception');
+            $oDTValidateRequestResponse
+                ->set_bSuccess(false)
+                ->add_aMessage(
+                    DTValidateMessage::create()
+                        ->set_sSubject('Exception')
+                        ->set_sBody($oInvalidOptionException->getMessage())
+                );
+
+            return $oDTValidateRequestResponse;
+        }
+
+        // requirement: it has to be Psr7
+        $oPsrRequest = new PsrRequest($oDTRequestCurrent);
         $aValidationResult = $oOpenApiValidation->validateRequest(
             // PSR7 Request Object
             $oPsrRequest,